requirements.txt
pdfplumber>=0.10.0
python-docx>=1.0.0
rich>=13.0.0
